refactor chromecast logic out of component
