+++ frontend/components.d.ts
@@ -33,6 +33,7 @@ declare module 'vue' {
     HeroGenre: typeof import('./src/components/HeroGenre.vue')['default']
     HeroPodcast: typeof import('./src/components/HeroPodcast.vue')['default']
     HRule: typeof import('./src/components/HRule.vue')['default']
+    'IconFluentEmojiHighContrast:spiderWeb': typeof import('~icons/fluent-emoji-high-contrast/spider-web')['default']
     IconIonShuffleSharp: typeof import('~icons/ion/shuffle-sharp')['default']
     IconNrkChevronDown: typeof import('~icons/nrk/chevron-down')['default']
     IconNrkClock: typeof import('~icons/nrk/clock')['default']
