+++ frontend/route-map.d.ts
@@ -164,13 +164,6 @@ declare module 'vue-router/auto-routes' {
       Record<never, never>,
       | never
     >,
-    '/test': RouteRecordInfo<
-      '/test',
-      '/test',
-      Record<never, never>,
-      Record<never, never>,
-      | never
-    >,
     '/tracks/': RouteRecordInfo<
       '/tracks/',
       '/tracks',
@@ -352,14 +345,6 @@ declare module 'vue-router/auto-routes' {
     'src/pages/settings.vue': {
       routes:
         | '/settings'
-      views:
-        | never
-      pathParamNames:
-        | never
-    }
-    'src/pages/test.vue': {
-      routes:
-        | '/test'
       views:
         | never
       pathParamNames:
