--- frontend/src/logic/browser.ts
@@ -1,4 +1,0 @@
-export function isBrowserChrome(): boolean {
-  return /Chrome/.test(navigator.userAgent)
-    && /Google Inc/.test(navigator.vendor)
-}
