--- frontend/src/pages/test.vue
@@ -1,10 +1,0 @@
-<script setup lang="ts">
-import { isBrowserChrome } from '~/logic/browser'
-
-const isChrome = isBrowserChrome()
-</script>
-
-<template>
-  is Chrome?: {{ isChrome }}
-  <ChromecastButton />
-</template>
