--- frontend/types/chromecast.d.ts
@@ -1,3 +1,0 @@
-interface Window {
-  __onGCastApiAvailable?: (isAvailable: boolean) => void
-}
