+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "Self Hosted Music Server",
   "main": "index.js",
-  "packageManager": "pnpm@11.1.2",
+  "packageManager": "pnpm@11.5.3",
   "workspaces": [
     "./frontend"
   ],
